Separated CausalDAG, CausalSpecification, and Scenario
